Replace node-fetch by node:http
The mu-javascript-template is no longer compatible with node-fetch
because it transforms imports into require() which are not supported by
node-fetch. The only solution is to use an ancient version of node-fetch
from when require() was still supported, or use something else entirely.
Insead of hunting for a new library at the risk of having to deal
with similar issues, I opted to just use the built-in http library. This
is crude, but gets the job done and is very stable over time.
Sorry, not sorry.